requirements.txt
aiohttp==3.7.4.post0
click==8.0.3
discord.py==1.7.3
discord_components==2.1.2
emoji==1.6.1
pandas==1.3.4
pony==0.7.14
python-dotenv==0.19.2
requests==2.26.0
tqdm==4.62.3
